Remove the preview and put the builder back into the sidebar.
