+++ client/Sidebar/BlueprintPanel.jsx
@@ -1,45 +1,177 @@
-import { useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
+import { toast } from 'react-toastify';
 import { useAppState } from '../context/AppStateContext.jsx';
-import { BlueprintConfigOverlay } from '../BlueprintConfigOverlay.jsx';
+import { useBlueprintFormState } from '../state/useBlueprintFormState.js';
+import { errorMessage } from '../utils/actions.js';
+import { postJSON } from '../utils/api.js';
+import { EnvironmentSection } from '../blueprint/EnvironmentSection.jsx';
+import { SiteSettingsSection } from '../blueprint/SiteSettingsSection.jsx';
+import { SlugListSection } from '../blueprint/SlugListSection.jsx';
+import { ContentSection } from '../blueprint/ContentSection.jsx';
 import { SectionBadge } from './SectionBadge.jsx';
 
-function getBlueprintSummary(blueprint) {
-  if (!blueprint) return 'No blueprint loaded';
-  const plugins = (blueprint.steps ?? []).filter((s) => s.step === 'installPlugin').length;
-  const themes = (blueprint.steps ?? []).filter((s) => s.step === 'installTheme').length;
-  const wp = blueprint.preferredVersions?.wp;
-  const parts = [];
-  if (plugins > 0) parts.push(`${plugins} plugin${plugins === 1 ? '' : 's'}`);
-  if (themes > 0) parts.push(`${themes} theme${themes === 1 ? '' : 's'}`);
-  if (wp && wp !== 'latest') parts.push(`WP ${wp}`);
-  return parts.length > 0 ? parts.join(', ') : 'Default configuration';
-}
-
 export function BlueprintPanel() {
-  const { blueprint, isBlueprintModified } = useAppState();
-  const [overlayOpen, setOverlayOpen] = useState(false);
+  const { blueprint: appBlueprint, defaultBlueprint, setBlueprint, isBlueprintModified } = useAppState();
+  const { formState, updateForm, loadBlueprint, compiledBlueprint, hasExtraSteps } =
+    useBlueprintFormState(appBlueprint);
+
+  const [activeTab, setActiveTab] = useState('form');
+  const [jsonDraft, setJsonDraft] = useState(() => JSON.stringify(compiledBlueprint, null, 2));
+  const [jsonError, setJsonError] = useState('');
+  const [isSaving, setIsSaving] = useState(false);
+  const skipSyncRef = useRef(false);
+
+  useEffect(() => {
+    if (skipSyncRef.current) {
+      skipSyncRef.current = false;
+      return;
+    }
+    setJsonDraft(JSON.stringify(compiledBlueprint, null, 2));
+    setJsonError('');
+  }, [compiledBlueprint]);
+
+  function currentBlueprint() {
+    if (activeTab === 'json') {
+      try { return JSON.parse(jsonDraft); } catch { /* fall through */ }
+    }
+    return compiledBlueprint;
+  }
+
+  async function persist(bp) {
+    await postJSON('/api/save-blueprint', { blueprint: bp });
+    setBlueprint(bp);
+  }
+
+  async function handleSave() {
+    setIsSaving(true);
+    try {
+      await persist(currentBlueprint());
+    } catch (err) {
+      toast.error(errorMessage(err, 'Could not save blueprint'));
+    } finally {
+      setIsSaving(false);
+    }
+  }
+
+  function handleReset() {
+    if (!defaultBlueprint) return;
+    if (!window.confirm('Reset all fields to the default blueprint? Your current changes will be lost.')) return;
+    loadBlueprint(defaultBlueprint);
+  }
+
+  function switchTab(newTab) {
+    if (activeTab === 'json' && newTab === 'form') {
+      try {
+        const parsed = JSON.parse(jsonDraft);
+        skipSyncRef.current = true;
+        loadBlueprint(parsed);
+        setJsonError('');
+        setActiveTab(newTab);
+      } catch (err) {
+        setJsonError(errorMessage(err, 'Invalid JSON'));
+      }
+    } else {
+      setActiveTab(newTab);
+    }
+  }
 
   return (
-    <>
-      <details id="blueprint-section">
-        <summary>
-          <span>Environment / Blueprint</span>
-          <SectionBadge hidden={!isBlueprintModified}>custom</SectionBadge>
-        </summary>
-        <div className="blueprint-body">
-          <p className="blueprint-summary">{getBlueprintSummary(blueprint)}</p>
+    <details id="blueprint-section">
+      <summary>
+        <span>Environment / Blueprint</span>
+        <SectionBadge hidden={!isBlueprintModified}>custom</SectionBadge>
+      </summary>
+      <div className="blueprint-body">
+        <div className="blueprint-panel-actions">
           <button
-            className="configure-blueprint-btn"
             type="button"
-            onClick={() => setOverlayOpen(true)}
+            className="bp-action-btn bp-action-btn--ghost"
+            onClick={handleReset}
+            disabled={!defaultBlueprint}
           >
-            Configure Blueprint
+            Reset
+          </button>
+          <button
+            type="button"
+            className="bp-action-btn bp-action-btn--primary"
+            onClick={handleSave}
+            disabled={isSaving}
+          >
+            {isSaving ? 'Saving…' : 'Save'}
           </button>
         </div>
-      </details>
-      {overlayOpen && (
-        <BlueprintConfigOverlay onClose={() => setOverlayOpen(false)} />
-      )}
-    </>
+
+        <div className="blueprint-tabs" role="tablist">
+          <button
+            role="tab"
+            aria-selected={activeTab === 'form'}
+            aria-controls="blueprint-tab-form"
+            className={`blueprint-tab-btn${activeTab === 'form' ? ' active' : ''}`}
+            onClick={() => switchTab('form')}
+          >
+            Form
+          </button>
+          <button
+            role="tab"
+            aria-selected={activeTab === 'json'}
+            aria-controls="blueprint-tab-json"
+            className={`blueprint-tab-btn${activeTab === 'json' ? ' active' : ''}`}
+            onClick={() => switchTab('json')}
+          >
+            JSON
+          </button>
+        </div>
+
+        <div
+          id="blueprint-tab-form"
+          role="tabpanel"
+          hidden={activeTab !== 'form'}
+          className="blueprint-form-tab"
+        >
+          {hasExtraSteps && (
+            <div className="blueprint-notice" role="status">
+              This blueprint contains advanced steps that are only editable in the JSON tab. They are preserved in the compiled output.
+            </div>
+          )}
+          <EnvironmentSection formState={formState} updateForm={updateForm} />
+          <SiteSettingsSection formState={formState} updateForm={updateForm} />
+          <SlugListSection
+            title="Plugins" sectionId="section-plugins"
+            items={formState.plugins} onUpdate={(plugins) => updateForm({ plugins })}
+            itemType="plugin" inputId="bf-plugin-slug"
+            inputPlaceholder="WordPress.org plugin slug"
+          />
+          <SlugListSection
+            title="Themes" sectionId="section-themes"
+            items={formState.themes} onUpdate={(themes) => updateForm({ themes })}
+            itemType="theme" inputId="bf-theme-slug"
+            inputPlaceholder="WordPress.org theme slug"
+          />
+          <ContentSection formState={formState} updateForm={updateForm} />
+        </div>
+
+        <div
+          id="blueprint-tab-json"
+          role="tabpanel"
+          hidden={activeTab !== 'json'}
+          className="blueprint-json-tab"
+        >
+          <textarea
+            id="blueprint-preview"
+            spellCheck="false"
+            className={jsonError ? 'invalid' : ''}
+            value={jsonDraft}
+            onChange={(e) => setJsonDraft(e.target.value)}
+            aria-label="Blueprint JSON editor"
+            aria-describedby={jsonError ? 'blueprint-json-error' : undefined}
+          />
+          {jsonError && (
+            <p id="blueprint-json-error" className="blueprint-json-error">
+              {jsonError}
+            </p>
+          )}
+        </div>
+      </div>
+    </details>
   );
 }
